requirements.txt
torch
torchvision
numpy
pandas
matplotlib
scikit-learn
pillow
streamlit
jupyter
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
numpy>=1.24.0
jupyter>=1.0.0
ipywidgets>=8.0.0
streamlit>=1.28.0
Pillow>=10.0.0
tqdm>=4.65.0
